Переименовал EnableDataProtocol/Class в ValidateProtocol/Manager Убрал метод, который проверял данные и изменял UI - данный функционал размазал по презентеру и view
